Make caching provider not require file length
This is important because blobnet would require a second network round
trip to check the file length, which isn't good for workloads like
blobnet where we shouldn't have range issues regardless because we store
it in the image manifest.

To do this, the blobnet API is slightly modified so that ranges dangling
past the end of the file are now valid and truncated automatically. This
is consistent with S3's API. I also corrected for some of the error
handling logic.